+++ catalog/model/payment/duitku_bnc.php
@@ -1,16 +1,16 @@
 <?php
 
-class ModelPaymentDuitkuVamandiri extends Model {
+class ModelPaymentDuitkuBnc extends Model {
   
   public function getMethod($address, $total) {
     
-    $this->load->language('payment/duitku_vamandiri');
+    $this->load->language('payment/duitku_bnc');
 
-    $query = $this->db->query("SELECT * FROM " . DB_PREFIX . "zone_to_geo_zone WHERE geo_zone_id = '" . (int)$this->config->get('duitku_vamandiri_geo_zone_id') . "' AND country_id = '" . (int)$address['country_id'] . "' AND (zone_id = '" . (int)$address['zone_id'] . "' OR zone_id = '0')");
+    $query = $this->db->query("SELECT * FROM " . DB_PREFIX . "zone_to_geo_zone WHERE geo_zone_id = '" . (int)$this->config->get('duitku_bnc_geo_zone_id') . "' AND country_id = '" . (int)$address['country_id'] . "' AND (zone_id = '" . (int)$address['zone_id'] . "' OR zone_id = '0')");
 
-    if ($this->config->get('duitku_vamandiri_total') > 0 && $this->config->get('duitku_vamandiri_total') > $total) {
+    if ($this->config->get('duitku_bnc_total') > 0 && $this->config->get('duitku_bnc_total') > $total) {
       $status = false;
-    } elseif (!$this->config->get('duitku_vamandiri_geo_zone_id')) {
+    } elseif (!$this->config->get('duitku_bnc_geo_zone_id')) {
       $status = true;
     } elseif ($query->num_rows) {
       $status = true;
@@ -32,9 +32,10 @@ class ModelPaymentDuitkuVamandiri extends Model {
 
     if ($status) {
       $method_data = array(
-        'code'       => 'duitku_vamandiri',
-        'title'      => $this->config->get('duitku_vamandiri_display_name'),
-        'sort_order' => $this->config->get('duitku_vamandiri_sort_order'),
+        'code'       => 'duitku_bnc',
+        'title'      => $this->config->get('duitku_bnc_display_name'),
+      
+        'sort_order' => $this->config->get('duitku_bnc_sort_order'),
         'terms'    => ''
       );
     }